General refactoring
- Created package spring_configuration to move there all configuration classes and web filters.
