+++ impc_prod_tracker/rest-api/src/main/java/org/gentar/spring_configuration/RestConfiguration.java
@@ -1,4 +1,4 @@
-package org.gentar.basic_data;
+package org.gentar.spring_configuration;
 
 import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.context.annotation.Bean;
