+++ impc_prod_tracker/rest-api/src/main/java/org/gentar/spring_configuration/filters/CORSFilter.java
@@ -1,4 +1,4 @@
-package org.gentar.security;
+package org.gentar.spring_configuration.filters;
 
 import java.io.IOException;
 import javax.servlet.Filter;
